Removed earth-demo scene and added game modes
Stripped out all the superfluous demo-code. Added GameModes enumerator
and included switch in the advance time method. This method will need
to be refactored (or at least renamed) as it is legacy and no longer
serves the purpose for which it was intended. Essentially, the elapsed
time, needs to be used for physics and other animation such as parallax
scrolling when we get there.